Pull params into params.py and start adding flags.
